temperature project
The given sample excel sheet describes the variations in temperature(in
Celsius) on a day to day basis for the months of November,December, January
and February for the city of Jaipur. You have to perform the following tasks with
reference to the data in the given sheet.
(Access the temperature datasheet using the link below:
https://docs.google.com/spreadsheets/d/1Gamkq8pGS1VGir9t_SL8apd7aq93-A
HnshZtjvrJPT0/edit?usp=sharing )1. Represent the data in the given sheet into an appropriate NumPy array so
that you can perform the following actions on it.
2. Write the dimensions and shape of the NumPy array that you have
created.
3. Print the daily temperatures for the first week of each month.
4. Print the temperatures for Tuesday of each month.
5. Print only the maximum temperature for all the weekdays of Dec and Feb.
6. Print all the days along with the week number in November when the
minimum temperature was less than 8 degrees.
7. Print all the weeks in Dec and Jan where the maximum temperature has
crossed a threshold of 20 degrees.
8. Check if there are any absurd values present in the dataset(like some temp
which should not be present in the data)
9. What strategy would you use to handle such data points?
10. Find and print the indexes of all the outlier(unusual) values present in
the above dataset.
11.Replace the outliers with an appropriate value.
12. Find the average max temperature for the winter months in Jaipur.
13. Find the weekly min avg temp for the month of Dec in Jaipur
14. Find the overall avg temp for the months Dec and Jan
15. Find the least temp experienced by the city in the month of Dec and
Jan. Also print the exact date( Day/Week/Month) for the same.
16. Find the max temp in the month of Feb and return its
date(Day/Week/Month)
17. Find the days in the month of Nov where the max temp of the day
dropped below the avg temp of the month. As
18. Convert the above dataset into an array where the weeks of the same
month must be present in the same row, but belonging to different months
should come in a row either below or above the selected month.
19. The above data is appropriate for an audience who follow the metric
system of measurement. Create an array that holds the same data but
presented in Fahrenheit.
20. Sort the above data in descending order on the basis of weekly average
for the month of Dec.
21. Sort the temp of the first three days of each month in descending order
on the basis of overall average for the whole winter.
22. Create an array that stores the difference between the min and max
temp for each day in all the winter months.
23. Find and store the difference between the max temp of two consecutive
days for each month of winter season.
24. Find and store the difference between the minimum temp of two
consecutive days for each month of the winter season.
25. Create an array by combining the data present in arrays created in q.23and q.24, to store the difference between the min and max temp of each
day of all the months for the whole winter season, in a single array.